+++ dapp/src/components/layout.js
@@ -20,6 +20,7 @@ import AppFooter from './AppFooter'
 import MarketingFooter from './MarketingFooter'
 import { adjustLinkHref } from 'utils/utils'
 import { assetRootPath } from 'utils/image'
+import { burnTimer } from 'utils/constants'
 
 const UNISWAP_URL =
   'https://app.uniswap.org/#/swap?inputCurrency=0xdac17f958d2ee523a2206206994597c13d831ec7&outputCurrency=0x2A8e1E676Ec238d8A992307B495b45B3fEAa5e86'
@@ -60,6 +61,14 @@ const Layout = ({
       console.error('Error OUSD REBASE OPT IN: ', error)
     }
   }
+
+  const { pathname } = useRouter()
+  const burnPage = pathname === '/burn'
+  const stakePage = pathname === '/earn'
+  const stakes = useStoreState(StakeStore, (s) => s)
+  const showStakingBanner = dapp && !stakePage && stakes.stakes?.length
+
+  const notice = showStakingBanner || burnTimer().days >= 0
 
   return (
     <>
@@ -151,7 +160,68 @@ const Layout = ({
           </a>
         </div>
       </div>
-      <Banner dapp={dapp} />
+      {notice && (
+        <div
+          className={classnames(
+            `notice ${showStakingBanner ? 'staking pt-2' : 'pt-3'} ${
+              burnPage ? 'burn' : ''
+            } ${dapp ? '' : 'px-lg-5'} text-white text-center pb-3`,
+            {
+              dapp,
+            }
+          )}
+        >
+          <div
+            className={`container d-flex flex-column flex-md-row align-items-center ${
+              dapp ? '' : 'nav px-lg-5'
+            }`}
+          >
+            {showStakingBanner ? (
+              <>
+                <div className="d-flex flex-column mt-0 justify-content-center px-4 px-md-0 text-md-left">
+          <div className="title-text">
+            {fbt(
+              'Changes are coming to OGN staking.',
+              'Changes are coming to OGN staking.'
+            )}
+          </div>
+          <div className="text">
+            {fbt(
+              'Your existing stakes will not be impacted. Claim your OGN at the end of your staking period.',
+              'Your existing stakes will not be impacted. Claim your OGN at the end of your staking period.'
+            )}
+          </div>
+        </div>
+        <div className="btn btn-dark mt-2 ml-md-auto">
+          <Link href={adjustLinkHref('/earn')}>Legacy staking</Link>
+        </div>
+              </>
+            ) : burnPage ? (
+              <>{fbt('OGV airdrop is live!', 'Airdrop notice')}
+              <a
+                href={process.env.AIRDROP_URL}
+                target="_blank"
+                rel="noopener noreferrer"
+                className="btn btn-dark mt-3 mt-md-0 ml-md-auto"
+              >
+                Check eligibility
+              </a></>
+            ) : (
+              <>{fbt(
+                'Only ' +
+                  fbt.param('burn-days', burnTimer().days) +
+                  ' days left to claim your OGV before the burn',
+                'Burn notice'
+              )}
+              <Link href={adjustLinkHref('/burn')}>
+                <a className="btn btn-dark gradient2 mt-3 mt-md-0 ml-md-auto">
+                  OGV Burn
+                </a>
+              </Link></>
+            )}
+          </div>
+        </div>
+      )}
       <main className={classnames({ dapp, short, shorter, medium })}>
         {dapp && <div className="container">{children}</div>}
         {!dapp && children}
@@ -188,6 +258,21 @@ const Layout = ({
           max-width: 940px;
           padding-left: 0px;
           padding-right: 0px;
+        }
+
+        .title-text {
+          font-size: 18px;
+          font-weight: bold;
+          line-height: 1.75;
+          color: white;
+        }
+
+        .text {
+          opacity: 0.8;
+          color: white;
+          line-height: normal;
+          font-size: 14px;
+          max-width: 1000px;
         }
       `}</style>
     </>
